refactorizacion jonathan
