+++ src/Farmacia/MovimientosGUI.form
@@ -153,7 +153,7 @@
         </constraints>
         <properties>
           <foreground color="-14780644"/>
-          <text value="Tipo de Farmacia.M.Movimiento"/>
+          <text value="Tipo de Farmacia.Movimiento"/>
         </properties>
       </component>
       <component id="f2cf6" class="javax.swing.JLabel">
