+++ src/Farmacia/ProductoGUI.java
@@ -1,8 +1,6 @@
 package Farmacia;
 
 import Conexion.ConexionBD;
-import Farmacia.C.ProductosDAO;
-import Farmacia.M.Productos;
 
 import javax.swing.*;
 import javax.swing.table.DefaultTableModel;
